feat: implement tao specific functions
TICKET: WIN-4296
