+++ modules/abstract-substrate/src/abstractSubstrateCoin.ts
@@ -98,15 +98,7 @@ export class SubstrateCoin extends BaseCoin {
 
   /** @inheritDoc **/
   generateKeyPair(seed?: Buffer): KeyPair {
-    const keyPair = seed ? utils.keyPairFromSeed(new Uint8Array(seed)) : new SubstrateKeyPair();
-    const keys = keyPair.getKeys();
-    if (!keys.prv) {
-      throw new Error('Missing prv in key generation.');
-    }
-    return {
-      pub: keys.pub,
-      prv: keys.prv,
-    };
+    throw new Error('Method not implemented.');
   }
 
   /** @inheritDoc **/
