+++ modules/abstract-substrate/src/lib/keyPair.ts
@@ -17,7 +17,7 @@ import * as nacl from 'tweetnacl';
 const TYPE = 'ed25519';
 const keyring = new Keyring({ type: TYPE });
 
-export class KeyPair extends Ed25519KeyPair {
+export abstract class KeyPair extends Ed25519KeyPair {
   /**
    * Public constructor. By default, creates a key pair with a random master seed.
    *
@@ -46,10 +46,11 @@ export class KeyPair extends Ed25519KeyPair {
    * or substrate format used for westend (starts with 5)
    */
   getAddress(format: DotAddressFormat): string {
-    let encodedAddress = this.createPolkadotPair().address;
-    encodedAddress = keyring.encodeAddress(encodedAddress, format as number);
+    // let encodedAddress = this.createPolkadotPair().address;
+    // encodedAddress = keyring.encodeAddress(encodedAddress, format as number);
 
-    return encodedAddress;
+    // return encodedAddress;
+    throw new Error('Method not implemented.');
   }
 
   /** @inheritdoc */
