+++ modules/sdk-coin-tao/src/lib/index.ts
@@ -1,7 +1,7 @@
 import * as Utils from './utils';
 import * as Interface from './iface';
 
-export { KeyPair } from './keyPair';
+export { TaoKeyPair } from './keyPair';
 export { Transaction } from './transaction';
 export { TransactionBuilder } from './transactionBuilder';
 export { TransferBuilder } from './transferBuilder';
